Delete failing errors_test.go (#251)
The only test case in that file only tests if the constructor function exists and returns an error with a correctly formatted string. This format changed in 4ce38f3546eb88b7265acb8d02e2e8f37f5fc5eb but wasn't updated in the test. And everything besides the format is also tested by the go compiler at built time. We don't need a special test for that.